fix: R-28 平台依赖项集中登记 — 可移植层预备
4 项可移植性改进:
1. fxnlms_mimo.h: volatile long → volatile int freeze_lms
   (long 在 LP64 为 64-bit, int 跨平台均为 32-bit)
   InterlockedExchange 调用加点显式 (LONG*) 转换

2. 新建 include/os_port.h — OS 可移植层:
   - gf_sleep_ms(): Win32→Sleep, Linux→usleep
   - gf_log_timestamp(): Win32/Linux→time+ctime, 裸机→GFANC_NO_TIME 去除

3. main_realtime.c: Sleep(100)→gf_sleep_ms(100), time/ctime→gf_log_timestamp()

4. calibrate_feedback.c: Sleep(100)→gf_sleep_ms(100)

与 include/os_atomic.h 互补: 原子操作用 os_atomic, 系统调用用 os_port.
Phase-2 (ARM Linux): 零改动 (宏自动切换).
Phase-3 (裸机): -DGFANC_NO_TIME 去除时间依赖.

+++ include/fxnlms_mimo.h
@@ -13,7 +13,7 @@ typedef struct {
     int          x_hist_ptr; /* R-12: x_hist 环形写指针 */
     float        step_size;
     float        leak;
-    volatile long freeze_lms; /* 发散检测: 1=冻结梯度更新, 0=正常 */
+    volatile int freeze_lms;  /* 发散检测: 1=冻结 (R-28: int 跨平台32-bit, long在LP64为64) */
 } fxnlms_mimo_t;
 
 int  fxnlms_init(fxnlms_mimo_t *fx, int E, int S, int L,
